+++ src/main/java/cn/curatorjin/smsweapon/smsweaponmod/blocks/FireBlock.java
@@ -1,4 +1,4 @@
-package com.smsweapon.smsweaponmod.blocks;
+package cn.curatorjin.smsweapon.smsweaponmod.blocks;
 
 import net.minecraft.block.Block;
 import net.minecraft.block.material.Material;
