+++ src/main/java/cn/curatorjin/smsweapon/smsweaponmod/items/materials/impl/FireDust.java
@@ -1,4 +1,4 @@
-package com.smsweapon.smsweaponmod.items.materials.impl;
+package cn.curatorjin.smsweapon.smsweaponmod.items.materials.impl;
 
 import net.minecraft.creativetab.CreativeTabs;
 import net.minecraft.entity.player.EntityPlayer;
