--- src/main/java/com/smsweapon/smsweaponmod/items/weapons/Weapon.java
@@ -1,7 +1,0 @@
-package com.smsweapon.smsweaponmod.items.weapons;
-
-import com.smsweapon.smsweaponmod.items.moulds.Mould;
-
-public interface Weapon {
-    public abstract Mould decompose();
-}
